// src/config/enums.ts
export enum MainKeyboard {
  menu = '📖 Меню',
  cart = '🛒 Корзина',
  contacts = '📍 Контакты'
}

export enum CheckoutKeyboard {
  shareContact = 'Поделиться контактом',
  back = '⬅️ Вернуться в меню'
}

export enum CartUi {
  title = '🛒 Моя Корзина:',
  cart = '🛒 Корзина',
  cancel = '❌ Отменить',
  edit = '✏️ Изменить',
  checkout = '✅ Заверните',
  reduce = '➖',
  increace = '➕',
  delete = '❌',
  placeholder = 'Добавь сюда вкусненького 🤤',
  total = 'Общая сумма заказа:',
  back = '⬅️ Назад'
}

export enum CheckoutUi {
  checkoutCancel = '❌ Оформление зказка отменено',
  contactStep = 'Поделись своим контактом или напиши номер в сообщении',
  commentStep = 'Есть какие-то комментарии к заказу?',
  phoneAdded = '✅ Номер добавлен',
  submitStep = '🙃 Спасибо, мы это учтем.',
  cartTotalTitle = 'Давайте сверим заказ',
  phoneValidationMessage = 'Это не похоже на номер или контакт',
  orderSubmit = '✅ Заказ подтверждаю',
  orderCancel = '❌ Отменить',
  orderCancaled = '🗑️ Закаказ удален',
  successMessage = '✅ Заказ №{N} оформлен. Спасибо. В ближайшее время мы свяжемся с Вами 🤙'
}

export enum GeneralTexts {
  contacts = '<b>График работы:</b>\nПН-ПТ с 10:00 до 22:00\nСБ-ВС c 10:00 до 23:00\n<b>Тел:</b> +380687889955\n<b>Адрес:</b> Днепр, Бульвар Кобзаря 3/1',
  start = 'Привет 👋\nЯ бот для оформления заказа в кафе TerAmo. Я помогу сэкономить Вам время, сделайте заказ по пути в кафе и приходите на готовенькое 🍕'
}

export enum Geo {
  lat = 48.525781,
  lon = 35.016239
}